requirements.txt
agency-swarm
discord.py>=2.3.2
web3
requests
python-dotenv
setuptools 